feat: major refactor of library structure and content
