+++ biosimulator_processes/processes/comparator_process.py
@@ -1,10 +1,4 @@
-from typing import *
 from process_bigraph import Process
-
-
-def generate_composite_index(results: Dict[str, float]):
-    num_population = len(list(results.keys()))
-    return sum(list(results.values())) / num_population
 
 
 class ODEComparator(Process):
@@ -60,27 +54,7 @@ class ODEComparator(Process):
             for simulator_name, simulator_process in self.simulator_instances.items()}
 
     def inputs(self):
-        # TODO: will each simulator have all of the same names/vals for this?
-        """floating_species_type = {
-            simulator_name: {
-                species_id: {
-                    '_type': 'float',
-                    '_apply': 'set'}
-                for species_id in simulator_species_names
-            }
-            for simulator_name, simulator_species_names in self.floating_species_ids.items()
-        }
-
-        model_params_type = {
-            param_id: {
-                '_type': 'float',
-                '_apply': 'set'}
-            for param_id in self.model_parameters_list}
-
-        reactions_type = {
-            reaction_id: 'float'
-            for reaction_id in self.reaction_list}"""
-
+        # TODO: will each ode simulator have all of the same names/vals for this?
         input_schema = {
             simulator_name: {
                 'time': 'float',
@@ -99,13 +73,6 @@ class ODEComparator(Process):
         return input_schema
 
     def outputs(self):
-        """floating_species_type = {
-            species_id: {
-                '_type': 'float',
-                '_apply': 'set'}
-            for species_id in self.floating_species_list
-        }"""
-
         output_schema = {
             simulator_name: {
                 'time': 'float',
@@ -121,7 +88,7 @@ class ODEComparator(Process):
         return output_schema
 
     def update(self, state, interval):
-        # TODO: do this in parallel
+        # TODO: instantiate parallel subprocesses for the simulation run
         results = {
             simulator_name: simulator_process.update(state, interval)
             for simulator_name, simulator_process in self.simulator_instances.items()}
@@ -135,72 +102,3 @@ class ODEComparator(Process):
                 results[simulator]['validation_score'] = diff  # TODO: make this more fine-grained
 
         return results
-
-
-class ComparisonDocument:
-    """To be called 'behind-the-scenes' by the Comparison REST API"""
-
-    def __init__(self,
-                 simulators: List[str],
-                 duration: int,
-                 num_steps: int,
-                 model_filepath: str,
-                 framework_type='deterministic',
-                 target_parameter: Dict[str, Union[str, float]] = None):
-        self.simulators = simulators
-        self.composite = {}
-        self.framework_type = framework_type
-        context = 'concentrations' if 'deterministic' in self.framework_type else 'counts'
-        self.species_port_name = f'floating_species_{context}'
-        self.species_store = [f'floating_species_{context}_store']
-        self._populate_composition(model_filepath)
-        self._add_emitter()
-
-    def _add_emitter(self):  # TODO: How do we reference different nesting levels?
-        self.composite['emitter'] = {
-            '_type': 'step',
-            'address': 'local:ram-emitter',
-            'config': {
-                'emit': {
-                    self.species_port_name: 'tree[float]',
-                    'time': 'float'}
-            },
-            'inputs': {
-                self.species_port_name: self.species_store,
-                'time': ['time_store']}
-        }
-
-    def _populate_composition(self, model_filepath: str):
-        context = 'concentrations' if 'deterministic' in self.framework_type else 'counts'
-        for index, process in enumerate(self.simulators):
-            self._add_ode_process_schema(
-                process_name=process,
-                species_context=context,
-                i=index,
-                model={'model_source': model_filepath}
-            )
-
-    def _add_ode_process_schema(
-            self,
-            process_name: str,
-            species_context: str,
-            i: int,
-            **config
-    ) -> None:
-        species_port_name = f'floating_species_{species_context}'
-        species_store = [f'floating_species_{species_context}_store']
-        self.composite[f'{process_name}_{i}'] = {
-            '_type': 'process',
-            'address': f'local:{process_name}',
-            'config': config,
-            'inputs': {
-                species_port_name: species_store,
-                'model_parameters': ['model_parameters_store'],
-                'time': ['time_store'],
-                'reactions': ['reactions_store']
-            },
-            'outputs': {
-                species_port_name: species_store,
-                'time': ['time_store']
-            }
-        }
